Fixes #2170 - Refactored some errorMessages imports
